requirements.txt
requests
json
pandas
matplotlib.pyplot
numpy
streamlit